making the input better
